feat: remove package-lock, rename to camel case
